Añadir a "[01-4] Guia 02- POO I" el Problema 4 Añadir Problema4.cpp Modificar ligeramente Problema3.cpp

+++ [01-4] Guia 02- POO I/Problema3.cpp
@@ -67,7 +67,7 @@ int main(int argc, char *argv[])
 
     precioDocena=GetValue("el precio por docena");
     costoTransporte=GetValue("el costo de transporte");
-    cantidadDocena=GetValue("la cantidad de docenas:");
+    cantidadDocena=(int)GetValue("la cantidad de docenas:");
 
     Compra objCompra(precioDocena,costoTransporte,cantidadDocena,descuento);
